+++ src/domain/auth/controller/oauth.controller.ts
@@ -1,21 +1,22 @@
-import { Body, Controller, HttpStatus, Post } from '@nestjs/common';
+import { Body, Controller, HttpStatus, Post, UseGuards } from '@nestjs/common';
 import { ResponseDto } from 'src/global/dtos/response.dto';
 import { OAuthLoginRequestDto } from '../dto/oauth-login.requst.dto';
 import { OAuthService } from '../service/oauth.service';
 import { OAuthSignUpRequestDto } from '../dto/oauth-signup.request.dto';
+import { JwtAuthUser, User } from 'src/global/decorators/jwt.decorator';
 
-@Controller('/api/oauth')
+@Controller('/api')
 export class OAuthController {
   constructor(private oauthService: OAuthService) {}
 
-  @Post('/kakao')
+  @Post('/oauth/kakao')
   async kakaoLogin(@Body() requestBody: OAuthLoginRequestDto) {
     const { idToken } = requestBody;
     const result = await this.oauthService.oauthKakaoLogin(idToken);
     return ResponseDto.okWithData(HttpStatus.OK, 'OAuth Login', result);
   }
 
-  @Post('/kakao/sign-up')
+  @Post('/oauth/kakao/sign-up')
   async kakaoSignUp(@Body() requestBody: OAuthSignUpRequestDto) {
     const result = await this.oauthService.kakaoOauthSignup(requestBody);
     return ResponseDto.okWithData(
